requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.17
decorator==4.4.1
Django==2.2
django-celery==3.2.1
django-filter==2.2.0
django-formfieldset==1.1
django-formtools==2.1
django-reversion==2.0.10
djangorestframework==3.11.0
elaphe3==0.2.0
ipython==5.5.0
ipython-genutils==0.2.0
kombu==3.0.37
mod-wsgi==4.6.8
numpy==1.18.0
olefile==0.46
pexpect==4.7.0
pickleshare==0.7.5
Pillow==4.3.0
pkg-resources==0.0.0
prompt-toolkit==1.0.18
psycopg2==2.7.3.1
ptyprocess==0.6.0
Pygments==2.5.2
python-dateutil==2.6.1
python-memcached==1.58
pytz==2017.2
redis==2.10.6
simplegeneric==0.8.1
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
xlrd==1.1.0
zbar-py==1.0.4
